+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fix ManOWar Air Supply being different when spawned in on land
   - Normal Conditions ManOWar will die on land after 5 minutes
   - If the biome is marked as a dry biome the ManOWar will die sooner (random between 1 and 5 minutes) (BWG Biomes Only)
+- Full Serene Seasons Compatability (Version 9.1.0.1 or newer)
 
 # 1.5.5
 - Add Crafting Table Block and Item Tags